+++ CHANGELOG.md
@@ -28,6 +28,8 @@ Changelog
  * removed `CannotEnableBindingException` and `CannotDisableBindingException`
  * replaced `BindingState::TYPE_NOT_LOADED` by `TYPE_NOT_FOUND` and
    `TYPE_NOT_ENABLED`
+ * `RepositoryManager::addPathMapping()` now throws an exception if the same
+   repository path is already mapped in the root package
    
 * 1.0.0-beta3 (2015-03-19)